backend/pyproject.toml
[project]
name = "winbid-ai-backend"
version = "0.1.0"
description = "Backend scaffold for WinBid-AI automated bid writing system"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "alembic>=1.13.1",
  "fastapi>=0.115.0",
  "openpyxl>=3.1.5",
  "openai>=1.51.0",
  "pillow>=10.4.0",
  "psycopg[binary]>=3.1.19",
  "pymupdf>=1.24.10",
  "pytesseract>=0.3.13",
  "uvicorn[standard]>=0.30.0",
  "xlrd>=2.0.1",
  "pydantic-settings>=2.3.0",
  "sqlalchemy>=2.0.30",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
